Declutter the repo
Removed Grunt and Node.JS stuff I am no longer doing any js work on the
website now that I use an off-the-shelf template. The only thing it was
kept for was minimizing the images.
Added a Makefile to manage some tasks and use some CLI tools to minimize
images.